src: rename files & session credential
