+++ lib/bearer_token_credential.js
@@ -1,5 +1,5 @@
 'use strict';
-const DefaultCredential = require('./defaultCredential');
+const DefaultCredential = require('./default_credential');
 
 
 class BearerTokenCredential extends DefaultCredential {
