+++ lib/credentials.js
@@ -1,13 +1,13 @@
 /* eslint-disable indent */
 'use strict';
 
-const AccessKeyCredential = require('./accessKeyCredential');
-const StsTokenCredential = require('./stsTokenCredential');
-const EcsRamRoleCredential = require('./ecsRamRoleCredential');
-const RamRoleArnCredential = require('./ramRoleArnCredential');
-const RsaKeyPairCredential = require('./rsaKeyPairCredential');
-const BearerTokenCredential = require('./bearerTokenCredential');
-const DefaultProvider = require('./provider/providerChan');
+const AccessKeyCredential = require('./access_key_credential');
+const StsTokenCredential = require('./sts_token_credential');
+const EcsRamRoleCredential = require('./ecs_ram_role_credential');
+const RamRoleArnCredential = require('./ram_role_arn_credential');
+const RsaKeyPairCredential = require('./rsa_key_pair_credential');
+const BearerTokenCredential = require('./bearer_token_credential');
+const DefaultProvider = require('./provider/provider_chan');
 
 class Credentials {
   constructor(config, runtime = {}) {
