+++ lib/default_credential.js
