+++ lib/provider/environment_variable_credentials_provider.js
@@ -1,5 +1,5 @@
 'use strict';
-const AccessKeyCredential = require('../accessKeyCredential');
+const AccessKeyCredential = require('../access_key_credential');
 class EnvironmentVariableCredentialsProvider {
   getCredential() {
     const accessKeyId = process.env.ALIBABA_CLOUD_ACCESS_KEY_ID;
