+++ lib/provider/instance_ram_role_credentials_provider.js
@@ -1,5 +1,5 @@
 'use strict';
-const EcsRamRoleCredential = require('../ecsRamRoleCredential');
+const EcsRamRoleCredential = require('../ecs_ram_role_credential');
 
 class InstanceRamRoleCredentialsProvider {
   getCredential() {
