+++ lib/provider/profile_credentials_provider.js
@@ -1,10 +1,10 @@
 'use strict';
-const AccessKeyCredential = require('../accessKeyCredential');
-const StsTokenCredential = require('../stsTokenCredential');
-const EcsRamRoleCredential = require('../ecsRamRoleCredential');
-const RamRoleArnCredential = require('../ramRoleArnCredential');
-const RsaKeyPairCredential = require('../rsaKeyPairCredential');
-const BearerTokenCredential = require('../bearerTokenCredential');
+const AccessKeyCredential = require('../access_key_credential');
+const StsTokenCredential = require('../sts_token_credential');
+const EcsRamRoleCredential = require('../ecs_ram_role_credential');
+const RamRoleArnCredential = require('../ram_role_arn_credential');
+const RsaKeyPairCredential = require('../rsa_key_pair_credential');
+const BearerTokenCredential = require('../bearer_token_credential');
 const utils = require('../util/utils');
 const fs = require('fs');
 const defaultPath = '~/.alibabacloud/credentials';
