+++ lib/provider/provider_chan.js
@@ -1,7 +1,7 @@
 'use strict';
-const EnvironmentVariableCredentialsProvider = require('./environmentVariableCredentialsProvider');
-const ProfileCredentialsProvider = require('./profileCredentialsProvider');
-const InstanceRamRoleCredentialsProvider = require('./instanceRamRoleCredentialsProvider');
+const EnvironmentVariableCredentialsProvider = require('./environment_variable_credentials_provider');
+const ProfileCredentialsProvider = require('./profile_credentials_provider');
+const InstanceRamRoleCredentialsProvider = require('./instance_ram_role_credentials_provider');
 const defaultProviders = [EnvironmentVariableCredentialsProvider, ProfileCredentialsProvider, InstanceRamRoleCredentialsProvider];
 
 const providerChan = {
