+++ lib/ram_role_arn_credential.js
@@ -1,18 +1,22 @@
 'use strict';
-const defaultCredential = require('./defaultCredential');
+const defaultCredential = require('./default_credential');
 const http = require('./util/http');
 const utils = require('./util/utils');
-const securityCredURL = 'http://100.100.100.200/latest/meta-data/ram/security-credentials/';
 
 
-class EcsRamRoleCredential extends defaultCredential {
-  constructor(role_name, runtime) {
+class RamRoleArnCredential extends defaultCredential {
+  constructor(config, runtime) {
     super({
-      type: 'ecs_ram_role',
-      role_name,
+      type: 'ram_role_arn',
+      access_key_id: config.access_key_id,
+      access_key_secret: config.access_key_secret,
+      role_arn: config.role_arn
     });
-    this.roleName = role_name;
-    this.runtime = runtime;
+    this.roleArn = config.role_arn;
+    this.policy = config.policy || false;
+    this.durationSeconds = config.role_session_expiration || 3600;
+    this.roleSessionName = config.role_session_name || 'role_session_name';
+    this.runtime = runtime || {};
     this.sessionCredential = null;
   }
   async getAccessKeyId() {
@@ -34,8 +38,14 @@ class EcsRamRoleCredential extends defaultCredential {
     return this.sessionCredential.SecurityToken;
   }
   async updateCredential() {
-    let url = securityCredURL + this.roleName;
-    let json = await http.request(url, {}, {}, this.accessKeySecret);
+    let json = await http.request('https://sts.aliyuncs.com', {
+      accessKeyId: this.accessKeyId,
+      roleArn: this.roleArn,
+      action: 'AssumeRole',
+      policy: this.policy,
+      DurationSeconds: this.durationSeconds,
+      roleSessionName: this.roleSessionName,
+    }, this.runtime, this.accessKeySecret);
     this.sessionCredential = json.Credentials;
   }
   needUpdateCredential() {
@@ -50,9 +60,11 @@ class EcsRamRoleCredential extends defaultCredential {
   }
   get configParams() {
     return {
-      roleName: 'role_name'
+      accessKeyId: 'access_key_id',
+      accessKeySecret: 'access_key_secret',
+      roleArn: 'role_arn'
     };
   }
 }
 
-module.exports = EcsRamRoleCredential;
+module.exports = RamRoleArnCredential;
