+++ lib/rsa_key_pair_credential.js
@@ -1,5 +1,5 @@
 'use strict';
-const defaultCredential = require('./defaultCredential');
+const defaultCredential = require('./default_credential');
 const fs = require('fs');
 const utils = require('./util/utils');
 const http = require('./util/http');
