+++ lib/sts_token_credential.js
@@ -1,5 +1,5 @@
 'use strict';
-const defaultCredential = require('./defaultCredential');
+const defaultCredential = require('./default_credential');
 
 
 class StsTokenCredential extends defaultCredential {
