+++ test/access_key_credential.test.js
