+++ test/bearer_token_credential.test.js
