+++ test/ecs_ram_role_credential.test.js
