+++ test/ram_role_arn_credential.test.js
