+++ test/sts_token_credential.test.js
